Reorganize Kubernetes upgrade files  (#710)
Reorganize Kubernetes upgrade files into a separate package

* Revert "Reorganize Kubernetes upgrade files"

This reverts commit 2f6732e3e66bdd0e36f2a4338ab687424ee06e36.
