+++ .gitignore
@@ -10,3 +10,4 @@ test/acs-engine-test/acs-engine-test
 
 test/acs-engine-test/acs-engine-test.exe
 pkg/operations/junit.xml
+pkg/operations/kubernetesupgrade/junit.xml
