+++ pkg/operations/kubernetesupgrade/upgradeworkflow.go
@@ -1,4 +1,4 @@
-package operations
+package kubernetesupgrade
 
 // UpgradeWorkFlow outlines various individual high level steps
 // that need to be run (one or more times) in the upgrade workflow.
