+++ pkg/operations/kubernetesupgrade/v162upgrader.go
@@ -1,4 +1,4 @@
-package operations
+package kubernetesupgrade
 
 import (
 	"encoding/json"
